+++ clib/Cargo.toml
@@ -18,10 +18,11 @@ log = "0.4.17"
 fuse-backend-rs = "0.10.0"
 nydus-api = { version = "0.2", path = "../api" }
 nydus-rafs = { version = "0.2.0", path = "../rafs" }
+nydus-storage = { version = "0.6.0", path = "../storage" }
 
 [features]
-baekend-s3 = ["nydus-rafs/backend-s3"]
-backend-oss = ["nydus-rafs/backend-oss"]
-backend-registry = ["nydus-rafs/backend-registry"]
-backend-http-proxy = ["nydus-rafs/backend-http-proxy"]
-backend-localdisk = ["nydus-rafs/backend-localdisk"]
+baekend-s3 = ["nydus-storage/backend-s3"]
+backend-oss = ["nydus-storage/backend-oss"]
+backend-registry = ["nydus-storage/backend-registry"]
+backend-http-proxy = ["nydus-storage/backend-http-proxy"]
+backend-localdisk = ["nydus-storage/backend-localdisk"]
