+++ rafs/Cargo.toml
@@ -39,11 +39,6 @@ assert_matches = "1.5.0"
 fusedev = ["fuse-backend-rs/fusedev"]
 virtio-fs = ["fuse-backend-rs/virtiofs", "vm-memory/backend-mmap"]
 vhost-user-fs = ["fuse-backend-rs/vhost-user-fs"]
-backend-oss = ["nydus-storage/backend-oss"]
-backend-s3 = ["nydus-storage/backend-s3"]
-backend-registry = ["nydus-storage/backend-registry"]
-backend-http-proxy = ["nydus-storage/backend-http-proxy"]
-backend-localdisk = ["nydus-storage/backend-localdisk"]
 
 [package.metadata.docs.rs]
 all-features = true
