+++ service/Cargo.toml
@@ -1,0 +1,50 @@
+[package]
+name = "nydus-service"
+version = "0.1.0"
+description = "Nydus Image Service Manager"
+authors = ["The Nydus Developers"]
+license = "Apache-2.0"
+homepage = "https://nydus.dev/"
+repository = "https://github.com/dragonflyoss/image-service"
+edition = "2018"
+resolver = "2"
+
+[dependencies]
+fuse-backend-rs = "0.10.1"
+libc = "0.2"
+log = "0.4.8"
+mio = { version = "0.8", features = ["os-poll", "os-ext"] }
+nix = "0.24.0"
+rust-fsm = "0.6.0"
+serde = { version = "1.0.110", features = ["serde_derive", "rc"] }
+serde_json = "1.0.51"
+thiserror = "1.0"
+time = { version = "0.3.14", features = ["serde-human-readable"] }
+tokio = { version = "1.24", features = ["macros"] }
+
+nydus-api = { version = "0.2.0", path = "../api", features = ["handler"] }
+nydus-error = { version = "0.2.3", path = "../error" }
+nydus-rafs = { version = "0.2.0", path = "../rafs" }
+nydus-storage = { version = "0.6.0", path = "../storage" }
+nydus-utils = { version = "0.4.0", path = "../utils" }
+
+vhost = { version = "0.5.0", features = ["vhost-user-slave"], optional = true }
+vhost-user-backend = { version = "0.7.0", optional = true }
+virtio-bindings = { version = "0.1", features = ["virtio-v5_0_0"], optional = true }
+virtio-queue = { version = "0.6.0", optional = true }
+vm-memory = { version = "0.9.0", features = ["backend-mmap"], optional = true }
+
+[dev-dependencies]
+vmm-sys-util = "0.10.0"
+
+[features]
+default = ["fuse-backend-rs/fusedev"]
+virtiofs = [
+    "fuse-backend-rs/vhost-user-fs",
+    "vm-memory",
+    "vhost",
+    "vhost-user-backend",
+    "virtio-queue",
+    "virtio-bindings",
+]
+
