+++ service/src/singleton.rs
@@ -11,7 +11,7 @@ use std::sync::{Arc, Mutex};
 
 use mio::Waker;
 use nydus_api::http::BlobCacheList;
-use nydus_app::BuildTimeInfo;
+use nydus_api::BuildTimeInfo;
 
 use crate::blob_cache::BlobCacheMgr;
 use crate::daemon::{
@@ -19,7 +19,7 @@ use crate::daemon::{
     NydusDaemon,
 };
 use crate::fs_service::FsService;
-use crate::{Error, Result, SubCmdArgs};
+use crate::{Error, Result, ServiceArgs};
 
 #[allow(dead_code)]
 struct ServiceController {
@@ -98,7 +98,11 @@ impl ServiceController {
 
 #[cfg(target_os = "linux")]
 impl ServiceController {
-    fn initialize_fscache_service(&self, subargs: &SubCmdArgs, path: &str) -> std::io::Result<()> {
+    fn initialize_fscache_service<T: ServiceArgs>(
+        &self,
+        args: &T,
+        path: &str,
+    ) -> std::io::Result<()> {
         // Validate --fscache option value is an existing directory.
         let p = match std::path::Path::new(&path).canonicalize() {
             Err(e) => {
@@ -120,9 +124,9 @@ impl ServiceController {
                 return Err(einval!("--fscache option contains invalid characters"));
             }
         };
-        let tag = subargs.value_of("fscache-tag").map(|s| s.as_str());
+        let tag = args.value_of("fscache-tag").map(|s| s.as_str());
 
-        let threads = if let Some(threads_value) = subargs.value_of("fscache-threads") {
+        let threads = if let Some(threads_value) = args.value_of("fscache-threads") {
             crate::validate_threads_configuration(threads_value).map_err(|err| einval!(err))?
         } else {
             1usize
@@ -234,16 +238,16 @@ impl DaemonStateMachineSubscriber for ServiceController {
 
 /// Create and start a Nydus daemon to host fscache and fusedev services.
 ///
-/// The `subargs` argument is derived from commandline options and controls services enabled.
-/// The fscache service will be enabled if `subargs` contains the `--fscache` argument.
-pub fn create_daemon(
-    subargs: &SubCmdArgs,
+/// The `args` argument is derived from commandline options and controls services enabled.
+/// The fscache service will be enabled if `args` contains the `--fscache` argument.
+pub fn create_daemon<T: ServiceArgs>(
+    args: &T,
     bti: BuildTimeInfo,
     waker: Arc<Waker>,
 ) -> std::io::Result<Arc<dyn NydusDaemon>> {
-    let id = subargs.value_of("id").map(|id| id.to_string());
-    let supervisor = subargs.value_of("supervisor").map(|s| s.to_string());
-    let config = match subargs.value_of("config") {
+    let id = args.value_of("id").map(|id| id.to_string());
+    let supervisor = args.value_of("supervisor").map(|s| s.to_string());
+    let config = match args.value_of("config") {
         None => None,
         Some(path) => {
             let config = std::fs::read_to_string(path)?;
@@ -273,8 +277,8 @@ pub fn create_daemon(
 
     service_controller.initialize_blob_cache(&config)?;
     #[cfg(target_os = "linux")]
-    if let Some(path) = subargs.value_of("fscache") {
-        service_controller.initialize_fscache_service(subargs, path)?;
+    if let Some(path) = args.value_of("fscache") {
+        service_controller.initialize_fscache_service(args, path)?;
     }
 
     let daemon = Arc::new(service_controller);
